+++ src/domain/App/index.ts
@@ -1,2 +1,3 @@
-export * from "./Store";
 export * from "./State";
+export * from "./Reducer";
+export * from "./Action";
