+++ src/index.tsx
@@ -1,2 +1,2 @@
-export * from "./client";
+export * from "./components";
 export * from "./server";
