+++ src/types/index.ts
@@ -1,5 +1,3 @@
-import * as React from "react";
-
 export interface Package {
   scripts: {
     [key: string]: string;
@@ -63,10 +61,3 @@ export interface Adjustment {
 export interface InitialProps {
   raw: Package;
 }
-
-export interface Action {
-  type: "UPDATE_KEY";
-  currentKey: string;
-}
-
-export type Dispatcher = React.Dispatch<Action>;
